Refactored from Nextjs `middleware` to `proxy` to address Vercel build errors.
